refactor: extract common types, interfaces into common package
