+++ packages/common/src/index.ts
@@ -5,4 +5,6 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-export { AxeResults, axeRuntimeExceptionMsgPrefix, axeVersion, getViolations } from './axe';
+export { A11yConfig, AxeResults, axeRuntimeExceptionMsgPrefix, axeVersion, getViolations } from './axe';
+export { AssertFunction, Options, WDIOBrowser } from './wdio';
+export { errMsgHeader, ExceptionList } from './format';
