+++ packages/preset-rules/src/index.ts
@@ -5,7 +5,7 @@
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
 
-export { getA11yConfig, A11yConfig } from './a11yConfig';
+export { getA11yConfig } from './a11yConfig';
 export { recommended } from './recommended';
 export { base } from './base';
 export { full } from './full';
