+++ packages/preset-rules/tsconfig.json
@@ -4,5 +4,6 @@
         "rootDir": "src",
         "outDir": "dist"
     },
-    "include": ["src"]
+    "include": ["src"],
+    "references": [{ "path": "../common" }]
 }
