+++ packages/wdio/src/setup.ts
@@ -4,7 +4,8 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { assertAccessible, assertAccessibleSync, WDIOBrowser } from './wdio';
+import { assertAccessible, assertAccessibleSync } from './wdio';
+import { WDIOBrowser } from '@sa11y/common';
 
 // https://webdriver.io/docs/typescript#adding-custom-commands
 declare global {
